Improve MetaInput and FontProvider

+++ runtime/src/de/fatox/meta/api/graphics/FontProvider.java
@@ -3,7 +3,7 @@ package de.fatox.meta.api.graphics;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 
 public interface FontProvider {
-    BitmapFont getFont(float size);
+    BitmapFont getFont(int size);
 
-    void write(float x, float y, String text, float size);
+    void write(float x, float y, String text, int size);
 }
